fix: raise when re-opening closed file
AIOFile.open() intended to raise an error if it was called after the
_file_obj had been closed. However in practice it could never happen, as
the function would always return None from the first if statement.

This is a change to the behaviour of open(), but it does appear to be
the intended behaviour, and it won't change the overall behaviour of
AIOFile, as even if someone was relying on open() succeeding after
close(), all methods that operate on the _file_obj would fail after
the re-open.